+++ packages/bpk-component-checkbox/src/BpkCheckbox.js
@@ -23,7 +23,6 @@ const BpkCheckbox = (props) => {
         disabled={disabled}
         {...rest}
       />
-      &nbsp;
       {label}
     </label>
   );