+++ packages/bpk-component-radio/src/BpkRadio.js
@@ -23,7 +23,6 @@ const BpkRadio = (props) => {
         disabled={disabled}
         {...rest}
       />
-      &nbsp;
       {label}
     </label>
   );