+++ changelog.md
@@ -5,6 +5,11 @@
 **Fixed:**
 - bpk-component-autosuggest:
   - Suggestion component can now recieve a className
+  
+- bpk-component-checkbox:
+- bpk-component-radio:
+- bpk-mixins:
+  - Checkboxes and radio buttons now display properly with multi line labels
 
 ## 2017-05-03 - Spinners come with less bloat